move private method to own file
